restructured directories and added jwt authentication basic classes
