+++ src/config/ROUTER/v1/ROUTER.js
@@ -1,11 +1,10 @@
 const express = require("express");
-const LOGGER = require("../../utils/LOGGER/v1/LOGGER");
+const LOGGER = require("../../LOGGER/v1/LOGGER");
 
-class ROUTER extends LOGGER {
+module.exports = class ROUTER extends LOGGER {
   constructor() {
     super(process.env.LOGS_FILE_PATH, process.env.LOG_LEVEL);
     this.router = express.Router();
   }
 }
 
-module.exports = ROUTER;
